pyproject.toml
[project]
name = "syrin"
version = "0.1.0"
description = "Python library for building AI agents with budget management, declarative definitions, and DSL-to-Python code generation. Companion to the Syrin DSL."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Syrin Labs" }]
keywords = ["ai", "agents", "llm", "budget", "dsl"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
Homepage = "https://syrin.ai/"
Repository = "https://github.com/syrin-labs/syrin-python"
dependencies = ["pydantic>=2.0", "httpx", "tiktoken", "python-dotenv>=1.2.1"]

[project.optional-dependencies]
litellm = ["litellm"]
anthropic = ["anthropic"]
openai = ["openai"]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
    "python-dotenv",
]

[project.scripts]
syrin = "Syrin.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/syrin"]

[tool.hatch.build.targets.sdist]
packages = ["src/syrin"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests", "examples"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",
    "E402",
    "ARG002",
    "ARG001",
    "F821",
    "F811",
] # line length, conditional imports, pytest fixtures, pre-existing test bugs

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "Syrin.providers.anthropic"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "Syrin.providers.openai"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "Syrin.providers.litellm"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "phoenix.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langfuse"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "qdrant_client"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "qdrant_client.models"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb.config"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psycopg2"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["src/Syrin"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
